Distributed ChatRoom App/requirements.txt
grpcio==1.49.1
grpcio-tools==1.49.1
PySimpleGUI==4.60.3
Pillow==9.2.0
numpy==1.23.3
six==1.16.0
tensorflow==2.10.0
tflearn==0.5.0
nltk==3.7
keyboard==0.13.5